@@ -1,6 +1,3 @@
-﻿using System;
-using Microsoft.AspNetCore.Razor.TagHelpers;
-
 namespace TagHelperPack;
 
 /// <summary>
@@ -27,15 +24,8 @@ public class IfTagHelper : TagHelper
     /// <inheritdoc />
     public override void Process(TagHelperContext context, TagHelperOutput output)
     {
-        if (context == null)
-        {
-            throw new ArgumentNullException(nameof(context));
-        }
-
-        if (output == null)
-        {
-            throw new ArgumentNullException(nameof(output));
-        }
+        ArgumentNullException.ThrowIfNull(context);
+        ArgumentNullException.ThrowIfNull(output);
 
         if (!Predicate)
         {
